+++ test/bigtest/mocks/jobs-logs.js
@@ -2,7 +2,10 @@ const jobsLogsData = [
   {
     id: '67dfac11-1caf-4470-9ad1-d533f6360bdd',
     hrId: '02',
-    jobProfileInfo: { name: 'Multilingual support check' },
+    jobProfileInfo: {
+      id: '22fafcc3-f582-493d-88b0-3c538480cd83',
+      name: 'Multilingual support check',
+    },
     fileName: 'import_28.mrc',
     progress: {
       jobExecutionId: '3d790a2a-86de-46f9-a9ae-c9109bba746b',
@@ -15,11 +18,15 @@ const jobsLogsData = [
     completedDate: '2018-11-11T14:10:34.000+0000',
     status: 'ERROR',
     uiStatus: 'ERROR',
+    userId: '12bbfdb5-91be-50ab-ad28-0d737fb5756a',
   },
   {
     id: '2e149aef-bb77-45aa-8a28-e139674b55e1',
     hrId: '03',
-    jobProfileInfo: { name: 'Standard BIB profile' },
+    jobProfileInfo: {
+      id: '32fafcc3-f582-493d-88b0-3c538480cd83',
+      name: 'Standard BIB profile',
+    },
     fileName: 'import_22.mrc',
     progress: {
       jobExecutionId: '2e149aef-bb77-45aa-8a28-e139674b55e1',
@@ -32,11 +39,15 @@ const jobsLogsData = [
     completedDate: '2018-11-05T14:21:57.000+0000',
     status: 'COMMITTED',
     uiStatus: 'RUNNING_COMPLETE',
+    userId: '18bbfdb5-91be-50ab-ad28-0d737fb5758a',
   },
   {
     id: '4aa3f7f9-3fe5-4a29-a149-72f7b08879da',
     hrId: '01',
-    jobProfileInfo: { name: 'BIB profile with customized Holdings' },
+    jobProfileInfo: {
+      id: '44fafcc3-f582-493d-88b0-3c538480cd83',
+      name: 'BIB profile with customized Holdings',
+    },
     fileName: 'import_22.mrc',
     progress: {
       jobExecutionId: '2e149aef-bb77-45aa-8a28-e139674b55e1',
@@ -49,6 +60,28 @@ const jobsLogsData = [
     completedDate: '2018-11-05T13:08:12.000+0000',
     status: 'COMMITTED',
     uiStatus: 'RUNNING_COMPLETE',
+    userId: '19bbfdb5-91be-50ab-ad28-0d737fb5759a',
+  },
+  {
+    id: '2e149aef-bb77-45aa-8a28-e139674b55e1',
+    hrId: '03',
+    jobProfileInfo: {
+      id: '32fafcc3-f582-493d-88b0-3c538480cd83',
+      name: 'Standard BIB profile',
+    },
+    fileName: 'import_21.mrc',
+    progress: {
+      jobExecutionId: '2e149aef-bb77-45aa-8a28-e139674b55e1',
+      total: 10,
+    },
+    runBy: {
+      firstName: 'Elliot',
+      lastName: 'Lane',
+    },
+    completedDate: '2018-11-05T14:22:57.000+0000',
+    status: 'COMMITTED',
+    uiStatus: 'RUNNING_COMPLETE',
+    userId: '18bbfdb5-91be-50ab-ad28-0d737fb5758a',
   },
 ];